safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.imageData.get.source.size.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.get.source.size</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var imgdata1 = ctx.getImageData(0, 0, 10, 10);
var imgdata2 = ctx.getImageData(0, 0, 20, 20);
_assert(imgdata2.width > imgdata1.width, "imgdata2.width > imgdata1.width");
_assert(imgdata2.height > imgdata1.height, "imgdata2.height > imgdata1.height");


});
</script>
