LayoutTests/canvas/philip/tests/2d.imageData.get.source.size-expected.txt

< [index] >
2d.imageData.get.source.size
getImageData() returns bigger ImageData for bigger source rectangle
References: 2d.imageData.get.basic
Actual output:
Passed
